# doc/api.yaml
openapi: 3.0.4
info:
  title: WASAtext
  description:  API Web Chat App, University project
  version: 1.1.1
servers:
  - url: https://api.web-chat-app.com/v1
    description: Web Chat Application API Description
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login
  schemas:  
...
    User:
      type: object
      description: Representation of a User in the system.
      properties:
        id:
          type: string
          format: uuid
          minLength: 36     
          maxLength: 36      
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Unique user id (UUID)
        username:
          type: string
          minLength: 3
          maxLength: 30
          example: steph_curry
          pattern: '^[a-zA-Z0-9._~-]+$'
          description: Unique username; only letters, numbers and underscore
        profilePhoto:
          type: string
          format: url                               # Url dell'immagine
          minLength: 10
          maxLength: 500
          pattern: '^https?:\/\/[^\s\/$.?#].[^\s]*$'
          description: URL User profile photo
        status:
          type: string 
          minLength: 1
          maxLength: 200
          pattern: '.*'               # gestire lato server
          description: User Status     
          example: "Available, write to me! :)"
      required:
        - id 
        - username
        - profilePhoto
...
    UpdateUsernameRequest:
      type: object
      description: Requested representation to change your username
      properties:
        newUsername:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9._~-]+$'
          description: New username (must be unique)
      required: 
        - newUsername
    Error:
      type: object
      description: Textual representation of the error
      properties:
        error:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '.*'             
          description: Error message
        code:
          type: string
          minLength: 1
          maxLength: 50
          pattern: '.*'             
          description: Error code
        details:
          type: string
          minLength: 0
          maxLength: 500
          pattern: '.*'             
          description: Additional error details
      required: 
        - error
  
## ENDPOINTS
paths:
...
  /wasatext/{username}/profile:

    get:                                                                    # Ottengo i dettagli del profilo che è loggato in questo momento
      operationId: myProfileDetails
      summary: Get current user profile
      description: Returns all the information about the user logged in to the session
      tags:
        - Users
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3                 
            maxLength: 30         
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: Username of the user loggin in this session
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "f4a8c2d1-0e9b-41b6-bd3c-527ef43ac821"
                username: "filippo_dev"
                profilePhoto: "https://web-chat-app.com/users/profile-images/filippo_dev.jpg"
  /wasatext/{username}/profile/status:

    patch:                                                                  # Modifico il profilo utente, dell'utente loggato in questo momento
      operationId: modifyProfileStatus
      summary: Update current user profile status
      description: Update current User profile logged in this session
      tags:
        - Users
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3                 
            maxLength: 30         
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: Username of the user loggin in this session
      requestBody:
        required: true
        content:
          application/json-patch+json:
            schema:
              type: object
              description: Update profile status 
              properties:
                status:
                  type: string
                  minLength: 0
                  maxLength: 200
                  pattern: '.*'
                  description: User status message
      responses:
        '200':
          description: Profile Status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "f4a8c2d1-0e9b-41b6-bd3c-527ef43ac821"
                username: "filippo_dev"
                profilePhoto: "https://web-chat-app.com/users/profile-images/filippo_dev_new.jpg"
  /wasatext/{username}/user_search:                                          # Ricercare un utente tra tutti gli utenti registrati nella web app
    get:
      operationId: searchUser
      summary: Search for a user by username
      description: Returns a list of users whose username contains the search string
      tags:
        - Users
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3                 
            maxLength: 30  
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: Username of the user loggin in this session
        - name: usernameSearched    # Con questo parametro di query vado a filtrare la lista 
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 30
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: Partial username to search for
      responses:
        '200':
          description: List of users found
          content:
            application/json:
              schema:
                type: object
                description: List of users whose username contains the search string 
                properties:
                  users:
                    type: array
                    minItems: 0
                    maxItems: 10000 
                    items:
                      $ref: "#/components/schemas/User"
                    description: Array of matching users. Empty array if no results.
              example:
                users:
                  - id: "f4a8c2d1-0e9b-41b6-bd3c-527ef43ac821"
                    username: "steph_curry"
                    profilePhoto: "https://web-chat-app.com/users/profile-images/steph_curry.png"
                  - id: "a3f6c4d1-0e2b-12b5-bd6c-527ef43ac821"
                    username: "steph_current"
                    profilePhoto: "https://web-chat-app.com/users/profile-images/steph_current.png"

        '404':
          description: User Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /wasatext/{username}/profile/username:
    put:                                               # Modifico il mio username
      operationId: setMyUserName
      summary: Update current user's username
      description: |-
        Updates the username of the currently authenticated user.  
        The request must include the new desired username in the request body.  
        If the username is valid and not already taken, it is updated successfully  
        and the updated user profile is returned.  
        If the username already exists, an error response is returned.
      tags:
        - Users
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3                 
            maxLength: 30
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: Username of the user loggin in this session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUsernameRequest"

      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "f4a8c2d1-0e9b-41b6-bd3c-527ef43ac821"
                username: "Ryan_Ted"
                profilePhoto: "https://web-chat-app.com/users/profile-images/ryan_ted.png"

        '400':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /wasatext/{username}/profile/photo:
    put:
      operationId: setMyPhoto
      summary: Update current user's profile photo
      description: |-
        Updates the profile photo of the currently authenticated user.
        The request must include the new photo URL in the request body.
        If the URL is valid, the profile photo is updated successfully
        and the updated user profile is returned.
      tags:
        - Users
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3                 
            maxLength: 30 
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: Username of the user loggin in this session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: profile photo
              properties:
                profilePhoto:
                  type: string
                  format: url
                  minLength: 10
                  maxLength: 500
                  pattern: '^https?:\/\/[^\s\/$.?#].[^\s]*$'
                  description: URL to new profile photo
              required:
                - profilePhoto
      responses:
        '200':
          description: Profile photo updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "f4a8c2d1-0e9b-41b6-bd3c-527ef43ac821"
                username: "filippo_dev"
                profilePhoto: "https://web-chat-app.com/users/profile-images/filippo_new_photo.png"
        '400':
          description: Invalid URL format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /wasatext/{username}/users/{usernameSearched}:
    get:                                                                      # ritornare un profilo utente tramite id
      operationId: getUserProfile
      summary: Get user profile by ID
      description: I return the user profile by passing his username as a parameter in the path
      tags:
        - Users
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3                 
            maxLength: 30    
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: Username of the user loggin in this session
        - name: usernameSearched
          in: path
          required: true
          schema:
            type: string
            minLength: 3                 
            maxLength: 30  
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: The username of the user to retrieve
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "f4a8c2d1-0e9b-41b6-bd3c-527ef43ac821"
                username: "filippo_dev"
                profilePhoto: "https://web-chat-app.com/users/profile-images/filippo_dev_new.jpg"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"